Use DateTime instead of Time::Piece.
Time::Piece problems appear to be responsible for a large number of test
failures.

In my testing, Time::Piece was giving me incorrect answers.

DateTime is a bit of a heavy beast, but it's already being loaded and used by
DateTime::Event:Easter anyway, so may as well use it.

These changes pass the existing tests.